Umfangreiches Refactoring
